+++ uu/id.go
@@ -3,10 +3,19 @@
 // that can be found in the LICENSE file.
 
 // Package uu provides type ID to keep, marshal and unmarshal UUID values.
+// See also RFC 4122.
 package uu
 
 import (
 	"fmt"
+)
+
+const (
+	// URNPrefix represents UUID URN prefix.
+	URNPrefix = "urn:uuid:"
+
+	// IDLength represents UUID length.
+	IDLength = 36
 )
 
 // ID represents UUID.
@@ -14,71 +23,79 @@ type ID struct {
 	Higher, Lower uint64
 }
 
-// MarshalText converts UUID to text.
-func (i ID) MarshalText() ([]byte, error) {
-	return []byte(i.String()), nil
+// Version returns UUID version.
+// Returned value has range 0-15.
+func (i ID) Version() int {
+	return int((i.Higher >> 12) & 0xf)
 }
 
-var (
-	starts = []int{0, 2, 4, 6, 9, 11, 14, 16, 19, 21, 24, 26, 28, 30, 32, 34}
-	digits = map[byte]uint64{
-		'0': 0,
-		'1': 1,
-		'2': 2,
-		'3': 3,
-		'4': 4,
-		'5': 5,
-		'6': 6,
-		'7': 7,
-		'8': 8,
-		'9': 9,
-		'A': 10,
-		'a': 10,
-		'B': 11,
-		'b': 11,
-		'C': 12,
-		'c': 12,
-		'D': 13,
-		'd': 13,
-		'E': 14,
-		'e': 14,
-		'F': 15,
-		'f': 15,
+// Variant returns UUID variant.
+// Returned value has range 0-3 (count of variant bits).
+func (i ID) Variant() int {
+	if i.Lower&0x8000000000000000 == 0 {
+		return 0
 	}
-)
+	if i.Lower&0x4000000000000000 == 0 {
+		return 1
+	}
+	if i.Lower&0x2000000000000000 == 0 {
+		return 2
+	}
+	return 3
+}
+
+// URN returns UUID in URN form.
+func (i ID) URN() string {
+	b, _ := DefaultFormatter([]byte("urn:uuid:"), i, 0)
+	return string(b)
+}
+
+// MarshalText converts UUID to text.
+func (i ID) MarshalText() ([]byte, error) {
+	b, err := Formatter(nil, i, 0)
+	if err != nil {
+		return nil, fmt.Errorf("uu.ID.MarshalText: %w", err)
+	}
+	return b, nil
+}
 
 // UnmarshalText parse UUID from text.
-func (i *ID) UnmarshalText(b []byte) error {
-	const length = 36
-	if l := len(b); l != length {
-		return fmt.Errorf("uu.ID.UnmarshalText: invalid length: expected %d instead of %d", length, l)
+func (i *ID) UnmarshalText(data []byte) error {
+	id, err := Parser(data, 0)
+	if err != nil {
+		return fmt.Errorf("uu.ID.UnmarshalText: %w", err)
 	}
-	if b[8] != '-' || b[13] != '-' || b[18] != '-' || b[23] != '-' {
-		return fmt.Errorf("uu.ID.UnmarshalText: %q: invalid format", b)
-	}
-	n := [2]uint64{}
-	for i, start := range starts {
-		for j := 0; j < 2; j++ {
-			if v, ok := digits[b[start+j]]; ok {
-				x := 124 - (i*2+j)*4
-				n[x>>6] |= v << (x & 0x3f)
-				continue
-			}
-			return fmt.Errorf("uu.ID.UnmarshalText: %q: invalid digit %c", b, b[start+j])
-		}
-	}
-	i.Higher = n[1]
-	i.Lower = n[0]
+	*i = id
 	return nil
+}
+
+// Format is implementation for fmt.Formatter.
+//
+//   ┌ Verb ┬ Format ───┬ Example ────────────────────────────────────────┐
+//   │ %s   │ Format(0) │ "00000000-0000-0000-0000-000000000000"          │
+//   │ %u   │ FormatURN │ "urn:uuid:00000000-0000-0000-0000-000000000000" │
+func (i ID) Format(f fmt.State, verb rune) {
+	f.Write(i.format(formatByVerb(verb)))
 }
 
 // String formats UUID for string output.
 func (i ID) String() string {
-	return fmt.Sprintf("%08x-%04x-%04x-%04x-%012x",
-		i.Higher>>32,
-		(i.Higher>>16)&0xffff,
-		i.Higher&0xffff,
-		i.Lower>>48,
-		i.Lower&0xffffffffffff,
-	)
+	return string(i.format(0))
+}
+
+func (i ID) format(f Format) []byte {
+	b, err := Formatter(nil, i, f)
+	if err != nil {
+		b, _ = DefaultFormatter(nil, i, f)
+	}
+	return b
+}
+
+func formatByVerb(verb rune) Format {
+	switch verb {
+	case 'u':
+		return FormatURN
+	default:
+		return 0
+	}
 }