+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## [Unreleased]
 
+## [0.7.0] - 2022-04-25
+### Added
+- Format/parse and `fmt.Formatter` support for `uu` package.
+- Function `uu.RandomID`.
+
+### Fixed
+- Error messages.
+
 ## [0.6.0] - 2022-04-19
 ### Added
 - Package `uu` with UUID support.
@@ -92,7 +100,8 @@
 ### Added
 - First release of util.
 
-[Unreleased]: https://github.com/livesport-tv/util/compare/v0.6.0...master
+[Unreleased]: https://github.com/livesport-tv/util/compare/v0.7.0...master
+[0.7.0]: https://github.com/livesport-tv/util/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/livesport-tv/util/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/livesport-tv/util/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/livesport-tv/util/compare/v0.3.0...v0.4.0